requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-MySQLdb==1.0.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient @ file:///C:/Users/Miguel/Documents/dev/pjts/flask/flask-mysql-crud/mysqlclient-1.4.6-cp38-cp38-win32.whl
python-dotenv==0.21.0
Werkzeug==2.2.2
zipp==3.8.1
